EZP-23969: slots: move to Common namespace
