+++ eZ/Publish/Core/Search/Common/Slot/DeleteLocation.php
@@ -7,15 +7,15 @@
  * @version //autogentag//
  */
 
-namespace eZ\Publish\Core\Search\Slot;
+namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
-use eZ\Publish\Core\Search\Slot;
+use eZ\Publish\Core\Search\Common\Slot;
 
 /**
- * A Search Engine slot handling TrashSignal.
+ * A Search Engine slot handling DeleteLocationSignal.
  */
-class Trash extends Slot
+class DeleteLocation extends Slot
 {
     /**
      * Receive the given $signal and react on it
@@ -24,10 +24,10 @@ class Trash extends Slot
      */
     public function receive( Signal $signal )
     {
-        if ( !$signal instanceof Signal\TrashService\TrashSignal )
+        if ( !$signal instanceof Signal\LocationService\DeleteLocationSignal )
             return;
 
-        $this->searchHandler->locationSearchHandler()->deleteLocation( $signal->locationId );
         $this->searchHandler->contentSearchHandler()->deleteLocation( $signal->locationId, $signal->contentId );
+        $this->searchHandler->locationSearchHandler()->deleteLocation( $signal->locationId );
     }
 }
