+++ eZ/Publish/Core/Search/Common/Slot/PublishVersion.php
@@ -7,15 +7,15 @@
  * @version //autogentag//
  */
 
-namespace eZ\Publish\Core\Search\Slot;
+namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
-use eZ\Publish\Core\Search\Slot;
+use eZ\Publish\Core\Search\Common\Slot;
 
 /**
- * A Search Engine slot handling SetContentStateSignal.
+ * A Search Engine slot handling PublishVersionSignal.
  */
-class SetContentState extends Slot
+class PublishVersion extends Slot
 {
     /**
      * Receive the given $signal and react on it
@@ -24,21 +24,14 @@ class SetContentState extends Slot
      */
     public function receive( Signal $signal )
     {
-        if ( !$signal instanceof Signal\ObjectStateService\SetContentStateSignal )
-        {
+        if ( !$signal instanceof Signal\ContentService\PublishVersionSignal )
             return;
-        }
-
-        $contentInfo = $this->persistenceHandler->contentHandler()->loadContentInfo( $signal->contentId );
 
         $this->searchHandler->contentSearchHandler()->indexContent(
-            $this->persistenceHandler->contentHandler()->load(
-                $contentInfo->id,
-                $contentInfo->currentVersionNo
-            )
+            $this->persistenceHandler->contentHandler()->load( $signal->contentId, $signal->versionNo )
         );
 
-        $locations = $this->persistenceHandler->locationHandler()->loadLocationsByContent( $contentInfo->id );
+        $locations = $this->persistenceHandler->locationHandler()->loadLocationsByContent( $signal->contentId );
         foreach ( $locations as $location )
         {
             $this->searchHandler->locationSearchHandler()->indexLocation( $location );
