+++ eZ/Publish/Core/Search/Common/Slot/Trash.php
@@ -7,15 +7,15 @@
  * @version //autogentag//
  */
 
-namespace eZ\Publish\Core\Search\Slot;
+namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
-use eZ\Publish\Core\Search\Slot;
+use eZ\Publish\Core\Search\Common\Slot;
 
 /**
- * A Search Engine slot handling DeleteContentSignal.
+ * A Search Engine slot handling TrashSignal.
  */
-class DeleteContent extends Slot
+class Trash extends Slot
 {
     /**
      * Receive the given $signal and react on it
@@ -24,13 +24,10 @@ class DeleteContent extends Slot
      */
     public function receive( Signal $signal )
     {
-        if ( !$signal instanceof Signal\ContentService\DeleteContentSignal )
+        if ( !$signal instanceof Signal\TrashService\TrashSignal )
             return;
 
-        // Delete Content
-        $this->searchHandler->contentSearchHandler()->deleteContent( $signal->contentId );
-
-        // Delete all Content locations
-        $this->searchHandler->locationSearchHandler()->deleteContent( $signal->contentId );
+        $this->searchHandler->locationSearchHandler()->deleteLocation( $signal->locationId );
+        $this->searchHandler->contentSearchHandler()->deleteLocation( $signal->locationId, $signal->contentId );
     }
 }
