--- eZ/Publish/Core/Search/Slot/UnhideLocation.php
@@ -1,40 +1,0 @@
-<?php
-/**
- * This file is part of the eZ Publish Kernel package
- *
- * @copyright Copyright (C) eZ Systems AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- * @version //autogentag//
- */
-
-namespace eZ\Publish\Core\Search\Slot;
-
-use eZ\Publish\Core\SignalSlot\Signal;
-use eZ\Publish\Core\Search\Slot;
-
-/**
- * A Search Engine slot handling UnhideLocationSignal.
- */
-class UnhideLocation extends Slot
-{
-    /**
-     * Receive the given $signal and react on it
-     *
-     * @param \eZ\Publish\Core\SignalSlot\Signal $signal
-     */
-    public function receive( Signal $signal )
-    {
-        if ( !$signal instanceof Signal\LocationService\UnhideLocationSignal )
-        {
-            return;
-        }
-
-        $this->searchHandler->contentSearchHandler()->indexContent(
-            $this->persistenceHandler->contentHandler()->load( $signal->contentId, $signal->currentVersionNo )
-        );
-
-        $this->searchHandler->locationSearchHandler()->indexLocation(
-            $this->persistenceHandler->locationHandler()->load( $signal->locationId )
-        );
-    }
-}
